Update regex to get the version
If there is any output while loading `.zshrc`, TypeProf will not start
because it cannot get the version.
In my use case, I was loading `.iterm2_shell_integration.zsh` in the
`.zshrc`.
refs: https://github.com/gnachman/iTerm2/blob/v3.5.13/Resources/shell_integration/iterm2_shell_integration.zsh

To avoid this problem, remove the `^` from the regular expression.
I think it is enough to check the end of the output with `$`.